pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fabagent-rag"
version = "0.1.0"
description = "一个使用 Milvus 作为向量数据库的 RAG 项目脚手架。"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "fastapi>=0.115.0",
    "openai>=1.54.0",
    "pymilvus>=2.6.12,<2.7",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.8.0",
]

[project.scripts]
rag = "fabagent_rag.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
